+++ pkg/label/label.go
@@ -1,13 +1,7 @@
 package label
 
 const (
-	Cluster = "giantswarm.io/cluster"
-
-	Organization = "giantswarm.io/organization"
-
-	ManagedBy = "giantswarm.io/managed-by"
-
-	// Labels, used in legacy cluster namespaces
+	// LegacyCustomer Labels, used in legacy cluster namespaces
 	LegacyCustomer = "customer"
 )