+++ service/controller/clusternamespace/clusternamespace_resources.go
@@ -8,7 +8,10 @@ import (
 	"github.com/giantswarm/operatorkit/v7/pkg/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/v7/pkg/resource/wrapper/retryresource"
 
+	"github.com/giantswarm/rbac-operator/service/controller/clusternamespace/resource/rbacappoperator"
+
 	"github.com/giantswarm/rbac-operator/service/controller/clusternamespace/resource/clusternamespaceresources"
+	"github.com/giantswarm/rbac-operator/service/controller/clusternamespace/resource/rbaccleaner"
 )
 
 type clusterNamespaceResourcesConfig struct {
@@ -32,8 +35,36 @@ func newClusterNamespaceResources(config clusterNamespaceResourcesConfig) ([]res
 		}
 	}
 
+	var rbacCleanerResource resource.Interface
+	{
+		c := rbaccleaner.Config{
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+		}
+
+		rbacCleanerResource, err = rbaccleaner.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var rbacAppOperatorResource resource.Interface
+	{
+		c := rbacappoperator.Config{
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+		}
+
+		rbacAppOperatorResource, err = rbacappoperator.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	resources := []resource.Interface{
 		clusterNamespaceResourcesResource,
+		rbacAppOperatorResource,
+		rbacCleanerResource,
 	}
 
 	{